merge new version

+++ src/core/service/api/storage/storage.js
@@ -1,20 +1,42 @@
+const STORAGE_DATA_TYPE = '__TYPE'
+const STORAGE_KEYS = 'uni-storage-keys'
+
+function parseValue (value) {
+  const types = ['object', 'string', 'number', 'boolean', 'undefined']
+  try {
+    const object = typeof value === 'string' ? JSON.parse(value) : value
+    const type = object.type
+    if (types.indexOf(type) >= 0) {
+      const keys = Object.keys(object)
+      // eslint-disable-next-line valid-typeof
+      if (keys.length === 2 && 'data' in object && typeof object.data === type) {
+        return object.data
+      } else if (keys.length === 1) {
+        return ''
+      }
+    }
+  } catch (error) { }
+}
+
 export function setStorage ({
   key,
   data
 } = {}) {
-  const value = {
-    type: typeof data === 'object' ? 'object' : 'string',
+  const type = typeof data
+  const value = type === 'string' ? data : JSON.stringify({
+    type,
     data: data
-  }
-  localStorage.setItem(key, JSON.stringify(value))
-  const keyList = localStorage.getItem('uni-storage-keys')
-  if (!keyList) {
-    localStorage.setItem('uni-storage-keys', JSON.stringify([key]))
-  } else {
-    const keys = JSON.parse(keyList)
-    if (keys.indexOf(key) < 0) {
-      keys.push(key)
-      localStorage.setItem('uni-storage-keys', JSON.stringify(keys))
+  })
+  try {
+    if (type === 'string' && parseValue(value) !== undefined) {
+      localStorage.setItem(key + STORAGE_DATA_TYPE, type)
+    } else {
+      localStorage.removeItem(key + STORAGE_DATA_TYPE)
+    }
+    localStorage.setItem(key, value)
+  } catch (error) {
+    return {
+      errMsg: `setStorage:fail ${error}`
     }
   }
   return {
@@ -32,13 +54,37 @@ export function setStorageSync (key, data) {
 export function getStorage ({
   key
 } = {}) {
-  const data = localStorage.getItem(key)
-  return data ? {
-    data: JSON.parse(data).data,
+  const value = localStorage && localStorage.getItem(key)
+  if (typeof value !== 'string') {
+    return {
+      data: '',
+      errMsg: 'getStorage:fail'
+    }
+  }
+  let data = value
+  const typeOrigin = localStorage.getItem(key + STORAGE_DATA_TYPE) || ''
+  const type = typeOrigin.toLowerCase()
+  if (type !== 'string' || (typeOrigin === 'String' && value === '{"type":"undefined"}')) {
+    try {
+      // 兼容H5和V3初期历史格式
+      let object = JSON.parse(value)
+      const result = parseValue(object)
+      if (result !== undefined) {
+        data = result
+      } else if (type) {
+        // 兼容App端历史格式
+        data = object
+        if (typeof object === 'string') {
+          object = JSON.parse(object)
+          // eslint-disable-next-line valid-typeof
+          data = typeof object === (type === 'null' ? 'object' : type) ? object : data
+        }
+      }
+    } catch (error) { }
+  }
+  return {
+    data,
     errMsg: 'getStorage:ok'
-  } : {
-    data: '',
-    errMsg: 'getStorage:fail'
   }
 }
 
@@ -52,14 +98,11 @@ export function getStorageSync (key) {
 export function removeStorage ({
   key
 } = {}) {
-  const keyList = localStorage.getItem('uni-storage-keys')
-  if (keyList) {
-    const keys = JSON.parse(keyList)
-    const index = keys.indexOf(key)
-    keys.splice(index, 1)
-    localStorage.setItem('uni-storage-keys', JSON.stringify(keys))
+  if (localStorage) {
+    // 兼容App端历史格式
+    localStorage.removeItem(key + STORAGE_DATA_TYPE)
+    localStorage.removeItem(key)
   }
-  localStorage.removeItem(key)
   return {
     errMsg: 'removeStorage:ok'
   }
@@ -72,7 +115,7 @@ export function removeStorageSync (key) {
 }
 
 export function clearStorage () {
-  localStorage.clear()
+  localStorage && localStorage.clear()
   return {
     errMsg: 'clearStorage:ok'
   }
@@ -82,18 +125,23 @@ export function clearStorageSync () {
   clearStorage()
 }
 
-export function getStorageInfo () { // TODO 暂时先不做大小的转换
-  const keyList = localStorage.getItem('uni-storage-keys')
-  return keyList ? {
-    keys: JSON.parse(keyList),
-    currentSize: 0,
-    limitSize: 0,
+export function getStorageInfo () {
+  const length = (localStorage && (localStorage.length || localStorage.getLength())) || 0
+  const keys = []
+  let currentSize = 0
+  for (let index = 0; index < length; index++) {
+    const key = localStorage.key(index)
+    if (key !== STORAGE_KEYS && key.indexOf(STORAGE_DATA_TYPE) + STORAGE_DATA_TYPE.length !== key.length) {
+      const value = localStorage.getItem(key)
+      currentSize += key.length + value.length
+      keys.push(key)
+    }
+  }
+  return {
+    keys,
+    currentSize: Math.ceil(currentSize * 2 / 1024),
+    limitSize: Number.MAX_VALUE,
     errMsg: 'getStorageInfo:ok'
-  } : {
-    keys: '',
-    currentSize: 0,
-    limitSize: 0,
-    errMsg: 'getStorageInfo:fail'
   }
 }
